refactor MonthsContext to tsx
